% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{analyze_eFast}
\alias{analyze_eFast}
\title{Analyze NetLogo simulation output of simdesign eFast}
\usage{
analyze_eFast(nl)
}
\arguments{
\item{nl}{nl object}
}
\description{
Analyze NetLogo simulation output of simdesign eFast
}
