% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{analyze_ff}
\alias{analyze_ff}
\title{Analyze NetLogo simulation output of simdesign full-factorial}
\usage{
analyze_ff(nl, funs)
}
\arguments{
\item{nl}{nl object}

\item{funs}{dplyr::funs list with the summary metrics one wants to have for the sensitivity results}
}
\description{
Analyze NetLogo simulation output of simdesign full-factorial
}
\keyword{internal}
