% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{analyze_sobol2007}
\alias{analyze_sobol2007}
\title{Analyze NetLogo simulation output of simdesign sobol2007}
\usage{
analyze_sobol2007(nl, metrics, funs)
}
\arguments{
\item{nl}{nl object}

\item{metrics}{vector of strings defining metric columns for evaluation. Defaults to metrics of the experiment within the nl object}

\item{funs}{dplyr::funs list with the summary metrics one wants to have for the sensitivity results}
}
\description{
Analyze NetLogo simulation output of simdesign sobol2007
}
\details{
The function calculates sobol sensitivity indices from the output results using the \link[sensitivity]{sensitivity} package.
}
\keyword{internal}
