% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_variables_constants.R
\name{check_constants_and_variables_logolink}
\alias{check_constants_and_variables_logolink}
\title{Backend function for eval_variables_constants (NetLogo 7+)}
\usage{
check_constants_and_variables_logolink(nl)
}
\arguments{
\item{nl}{nl object}
}
\value{
Character vector of detected issues
}
\description{
Checks constants and variable formatting of the experiment as required by Logolink
}
\details{
Checks constants and variable formatting of the experiment, as NetLogo 7+ (or at least Logolink)
has different formatting specifications than before.
}
\keyword{internal}
