% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_variables_constants.R
\name{check_value_format}
\alias{check_value_format}
\title{Helper function for \code{check_constants_and_variables_logolink()}}
\usage{
check_value_format(x, name = NULL)
}
\arguments{
\item{x}{value to check}

\item{name}{optional name of the checked entry}
}
\value{
Character vector of detected issues
}
\description{
Checks variables and constants for patterns problematic in NetLogo 7+
}
\keyword{internal}
