% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nl.R
\name{create_simulation_blocks}
\alias{create_simulation_blocks}
\title{Backend function for \code{run_nl_all()} (NetLogo 7+)}
\usage{
create_simulation_blocks(nl, block_size)
}
\arguments{
\item{nl}{See \code{run_nl_all()}}

\item{block_size}{See \code{run_nl_all()}}
}
\value{
A list of data frames, each representing one bundle of simulations
}
\description{
Splits \code{nl@simdesign@siminput} into blocks of size \code{block_size}, to be executed by \code{run_nl_block()}
}
\keyword{internal}
