% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_tibble.R
\name{list_to_tibble}
\alias{list_to_tibble}
\title{list_to_tibble}
\usage{
list_to_tibble(data_list)
}
\arguments{
\item{data_list}{a list of tibbles}
}
\value{
a combined tibble
}
\description{
Combines a list of tibbles to a tibble. If one or several columns are incompatible, the columns with the most abundant column type in the list are retained and the incompatible columns are set to NA.
}
