% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stuff.R
\name{load_model_parameters}
\alias{load_model_parameters}
\title{Report globals from a NetLogo model that is defined within a nl object}
\usage{
load_model_parameters(nl)
}
\arguments{
\item{nl}{nl object with a defined modelpath that points to a NetLogo model
(*.nlogo)}
}
\description{
Report globals from a NetLogo model that is defined within a nl
object
}
\details{
The function reads the NetLogo model file that is defined within the nl object
and reports all global parameters that are defined as widget elements on
the GUI of the NetLogo model.
Only globals that are found by this function are valid globals that can be
entered into the variables or constants vector of an experiment object.
}
\examples{
\dontrun{
load_model_parameters(nl)
}

}
