% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nl.R
\name{merge_result_blocks}
\alias{merge_result_blocks}
\title{Backend function for \code{run_nl_all()}}
\usage{
merge_result_blocks(nl, results_list)
}
\arguments{
\item{nl}{See \code{run_nl_all()}}

\item{results_list}{List of result tibbles returned by \code{run_nl_block()}}
}
\value{
Tibble with simulation output results
}
\description{
Merges results produced by \code{run_nl_block()} into one tibble. Temporarily handles column name restoration until \code{logolink} supports preserving original column names.
}
\keyword{internal}
