% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{nl_gensa}
\alias{nl_gensa}
\title{Wolf Sheep model sample data: gensa}
\format{
nl object with defined experiment, simdesign and model output
}
\usage{
nl_gensa
}
\description{
The dataset contains a complete nl object.
The nl object has been used to setup and run a simulated annealing simulation design.
}
\keyword{datasets}
