% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{nl_simple}
\alias{nl_simple}
\title{Wolf Sheep model sample data: simdesign simple}
\format{
nl object with defined experiment, simdesign and model output
}
\usage{
nl_simple
}
\description{
The dataset contains a complete nl object.
The nl object has been used to setup and run a simple simulation design.
It also contains the model outputs within the simdesign object.
}
\keyword{datasets}
