% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{nl_soboljansen}
\alias{nl_soboljansen}
\title{Wolf Sheep model sample data: simdesign soboljansen}
\format{
nl object with defined experiment, simdesign and model output
}
\usage{
nl_soboljansen
}
\description{
The dataset contains a complete nl object.
The nl object has been used to setup and run a soboljansen simulation design.
It also contains the model outputs within the simdesign object.
Further analysis output can be derived by submitting the dataset to analyze_nl().
}
\keyword{datasets}
