% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{nl_spatial}
\alias{nl_spatial}
\title{Wolf Sheep model sample data: spatial}
\format{nl object with defined experiment, simdesign and model output}
\usage{
nl_spatial
}
\description{
The dataset contains a complete nl object.
The nl object has been used to setup and run a simple simulation design.
For these simulations, spatial agent output has been collected.
metrics.turtles was set up to measure turtle coordinates.
metrics.patches was set up to measure colors of cells.
It also contains the model outputs within the simdesign object.
}
\keyword{datasets}
