% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_to_raster.R
\name{nl_to_raster}
\alias{nl_to_raster}
\title{Get spatial data from metrics.turtles and metrics.patches output}
\usage{
nl_to_raster(nl)
}
\arguments{
\item{nl}{nl object}

\item{coords}{nl object}
}
\value{
tibble with spatial data objects
}
\description{
Turn results from NetLogo in spatial data objects
}
\details{
Unnests output from run_nl into long format.
}
\examples{

}
