% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldoc_find_procedure_calls.R
\name{nldoc_find_procedure_calls}
\alias{nldoc_find_procedure_calls}
\title{Determine procedure calls}
\usage{
nldoc_find_procedure_calls(nlogocode)
}
\arguments{
\item{nlogocode}{vector of netlogo code strings}
}
\value{
tibble with procedure names and procedure calls
}
\description{
Determine procedure calls
}
\details{
The procedure searches netlogo code for procedure definitions and calls.
The information is stored within a tibble that can be further processed.
}
\keyword{internal}
