% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldoc_parse_modelcode.R
\name{nldoc_parse_modelcode}
\alias{nldoc_parse_modelcode}
\title{Parse model code}
\usage{
nldoc_parse_modelcode(nlogocode)
}
\arguments{
\item{nlogocode}{vector of netlogo code strings}
}
\value{
tibble with structured netlogo code
}
\description{
Parse model code
}
\details{
The procedure searches for noxygen commands within the NetLogo code.
This information is used to structure the netlogo code strings in a tibble.
Additionally, tibbles with gui and behavior space information are created.
}
\keyword{internal}
