% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldoc_read_nlogo.R
\name{nldoc_read_nlogo}
\alias{nldoc_read_nlogo}
\title{Read NetLogo model code from files}
\usage{
nldoc_read_nlogo(modelfiles)
}
\arguments{
\item{modelfiles}{vector of filepaths to model files}
}
\value{
vector of strings containing NetLogo model code
}
\description{
Read NetLogo model code from files
}
\details{
The procedure reads text from the provided model files and reports the code as a vector of strings.
}
\keyword{internal}
