% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldoc_table_bs.R
\name{nldoc_table_bs}
\alias{nldoc_table_bs}
\title{Read NetLogo behavior space experiments from files}
\usage{
nldoc_table_bs(modelfiles)
}
\arguments{
\item{modelfiles}{vector of filepaths to model files}
}
\value{
list containing NetLogo behavior space experiments
}
\description{
Read NetLogo behavior space experiments from files
}
\details{
The procedure reads text from the provided model files and reports a list of behavior space experiments.
}
\keyword{internal}
