% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldoc_table_gui.R
\name{nldoc_table_gui}
\alias{nldoc_table_gui}
\title{Read NetLogo GUI elements from files}
\usage{
nldoc_table_gui(modelfiles)
}
\arguments{
\item{modelfiles}{vector of filepaths to model files}
}
\value{
list containing NetLogo GUI elements
}
\description{
Read NetLogo GUI elements from files
}
\details{
The procedure reads text from the provided model files and reports a list of GUI elements.
}
\keyword{internal}
