% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldoc_write_nldoc.R
\name{nldoc_write_nldoc}
\alias{nldoc_write_nldoc}
\title{Write NetLogo documentation}
\usage{
nldoc_write_nldoc(
  noxygen,
  noxygen_it,
  noxygen_gui,
  noxygen_bs,
  outpath,
  output_format,
  number_sections,
  theme,
  date,
  toc
)
}
\arguments{
\item{noxygen}{list with parsed and processed noxygen tags from NetLogo model code}

\item{noxygen_it}{list with parsed and processed infotab strings}

\item{noxygen_gui}{list with parsed and processed noxygen tags from NetLogo GUI elements}

\item{noxygen_bs}{list with parsed and processed noxygen tags from NetLogo behavior space experiments}

\item{outpath}{Path to folder where rendered documentation should be created}

\item{output_format}{either "html", "pdf" or "docx"}

\item{number_sections}{TRUE/FALSE, if TRUE sections in the documentation will be numbered}

\item{theme}{markdown theme, supported themes are "journal", "cerulean", "flatly", "readable", "spacelab", "united", "cosmo"}

\item{date}{date that is printed in the documentation header}

\item{toc}{TRUE/FALSE, if TRUE the documentation contains a table of contents - only for html and pdf output format}
}
\value{
list containing NetLogo GUI elements
}
\description{
Write NetLogo documentation
}
\details{
The procedure uses parsed and processed noxygen tags to create and render a markdown documentation in the specified format.
}
\keyword{internal}
