% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlrx_functions.R
\name{nlrx_writeXML}
\alias{nlrx_writeXML}
\title{Write NetLogo BehaviorSpace Experiment File as xml}
\usage{
nlrx_writeXML(filename, experiment.name, n.repetitions, metrics.each.tick,
  setup.name, go.name, time.limit, metrics, parameters)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}

\item{cex}{character expansion for the text}

\item{mar}{margin paramaters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
None
}
\description{
Creates an xml file in NetLogo BehaviorSpace format, that can be used to start NetLogo simulations via command line.
}
\examples{
plot_crayons()

}
