% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_nl.R
\name{print.nl}
\alias{print.nl}
\title{Print content of nl object}
\usage{
\method{print}{nl}(x, ...)
}
\arguments{
\item{x}{nl object to print}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print content of nl object and embedded experiment and simdesign objects to console
}
\details{
Print content of the provided nl object in a readable format.
}
\examples{

print(nl_lhs)

}
