% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_r.R
\name{print.r}
\alias{print.r}
\title{Print content of r object}
\usage{
\method{print}{r}(x, ...)
}
\arguments{
\item{x}{r object to print}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print content of r object and embedded experiment and simdesign objects to console
}
\details{
Print content of the provided r object in a readable format.
}
\examples{

print(nl_lhs)

}
