% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constr.R
\name{r}
\alias{r}
\title{Construct a new r object
#'}
\usage{
r(modeltype = "used model type", obj_type = "r_obj", ...)
}
\arguments{
\item{modeltype}{A character string defining the modeltype}

\item{obj_type}{A character string defining the object (nl, R, ...)}

\item{...}{...}

\item{jvmmem}{Java virtual machine memory capacity in megabytes}

\item{experiment}{Holds a experiment S4 class object}

\item{simdesign}{Holds a simdesign S4 class object}
}
\value{
r S4 class object
}
\description{
Construct a new r object
}
\details{
r objects... TEXT ANPASSEN!
r objects are the main class objects used in the nlrx package.
These objects store all information that is needed to run NetLogo simulations.
nl objects are initialized with basic information on Netlogo and the model.

After setting up the nl object, an experiment needs to be attached.
The experiment class stores all information related to the NetLogo simulation experiment, such as runtime,
variables, constants, measurements, and more.

After attaching an experiment, different simdesign helper functions can be used to attach a simdesign to the nl object.
The simdesign helper functions use the variable definitions from the experiment within the nl object to generate a parameter tibble for simulations.
}
\examples{
outpath <- file.path("/home/out")


r <- r( modeltype = "used model type",
        obj_type = "r_obj",
        jvmmem = 1024)

}
