% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_all}
\alias{run_all}
\title{nl, später aber object}
\usage{
run_all(used_obj)
}
\arguments{
\item{nl}{nl object}
}
\value{
tibble with simulation output results
}
\description{
Execute simulations from R object
}
