% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nl.R
\name{run_nl_block}
\alias{run_nl_block}
\title{Backend function used by \code{run_nl_all()} and \code{run_nl_one()} to run simulations (NetLogo 7+)}
\usage{
run_nl_block(nl, block_df, block_number, threads)
}
\arguments{
\item{nl}{See \code{run_nl_all()}}

\item{block_df}{Data frame of simulation definitions, with one row being one run}

\item{block_number}{Assigned by \code{run_nl_all()}, used for naming the XML file.}
}
\value{
Tibble with simulation output results
}
\description{
Creates a BehaviorSpace XML file for simulations defined in \code{block_df} and executes them via \code{logolink}
}
\details{
Runs a single block (\code{block_df}) by writing a single BehaviorSpace XML file and executing it via \code{logolink}.
}
\keyword{internal}
