% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_r.R
\name{run_r_all}
\alias{run_r_all}
\title{Execute all R simulations from an R object}
\usage{
run_r_all(
  r,
  split = 1,
  cleanup.csv = TRUE,
  cleanup.xml = TRUE,
  cleanup.bat = TRUE
)
}
\arguments{
\item{r}{r object}

\item{split}{number of parts the job should be split into}

\item{cleanup.csv}{TRUE/FALSE, if TRUE temporary created csv output files will be deleted after gathering results.}

\item{cleanup.xml}{TRUE/FALSE, if TRUE temporary created xml output files will be deleted after gathering results.}

\item{cleanup.bat}{TRUE/FALSE, if TRUE temporary created bat/sh output files will be deleted after gathering results.}
}
\value{
tibble with simulation output results
}
\description{
Execute all R simulations from an R object with a defined experiment and simdesign
}
\details{
run_r_all executes all simulations ....
}
\examples{
\dontrun{

# Load nl object from test data:
nl <- nl_lhs

# Execute all simulations from an nl object with properly attached simdesign.
results <- run_r_all(nl)

# Run in parallel on local machine:
library(future)
plan(multisession)
results <- run_r_all(nl)

}
}
