% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_init.R
\name{simdesign-class}
\alias{simdesign-class}
\alias{.initClasses}
\title{S4 class \code{"simdesign"}}
\usage{
.initClasses()
}
\description{
empty

empty

empty
}
\section{Slots}{

\describe{
\item{\code{simmethod}}{A character string for method identification}

\item{\code{siminput}}{Input parameter tibble for model simulations}

\item{\code{simobject}}{List of model objects that are created for specific simulation designs and are needed for output analysis (e.g. sobol, morris, eFast)}

\item{\code{simseeds}}{A vector of model seeds that are used to repeatedly run simulations using the siminput data}

\item{\code{simoutput}}{Output simulation results can be attached to this slot as tibble}

\item{\code{expname}}{A character string defining the name of the experiment}

\item{\code{outpath}}{Path to a directory where experiment output will be stored}

\item{\code{repetition}}{A number which gives the number of repetitions for each row of the simulation design input tibble}

\item{\code{tickmetrics}}{Character string "true" runs defined metrics on each simulation tick. "false" runs metrics only after simulation is finished}

\item{\code{idsetup}}{character string or vector of character strings, defining the name of the NetLogo setup procedure}

\item{\code{idgo}}{character string or vector of character strings, defining the name of the NetLogo go procedure}

\item{\code{idfinal}}{character string or vector of character strings, defining the name of NetLogo procedures that should be run after the last tick}

\item{\code{runtime}}{number of model ticks that should be run for each simulation}

\item{\code{evalticks}}{vector of tick numbers defining when measurements are taken}

\item{\code{metrics}}{vector of strings defining valid NetLogo reporters that are taken as output measurements (e.g. c("count turtles", "count patches"))}

\item{\code{variables}}{a nested list of variables that are changed within a simulation design. The name of each sublist item has to be a valid global of the defined NetLogo model. Each list item consist of a min value, a max value, a step value and a qfun (e.g. list("paramA" = list(min=0, max=1, step=0.1, qfun="qunif")))}

\item{\code{constants}}{a list of constants that are kept constant within a simulation design. The name of each list item has to be a valid global of the defined NetLogo model (e.g. list("pNUM" = 12, "pLOGIC"="TRUE", "pSTRING"=""default""))}

\item{\code{simdesign}}{Holds a simdesign S4 class object}

\item{\code{nlversion}}{A character string defining the NetLogo version that is used}

\item{\code{nlpath}}{Path to the NetLogo main directory matching the defined version}

\item{\code{modelpath}}{Path to the NetLogo model file (*.nlogo) that is used for simulations}

\item{\code{jvmmem}}{Java virtual machine memory capacity in megabytes}

\item{\code{experiment}}{Holds a experiment S4 class object}
}}

