% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdesign_helper.R
\name{simdesign_soboljansen}
\alias{simdesign_soboljansen}
\title{Add a soboljansen simdesign to a nl object}
\usage{
simdesign_soboljansen(nl, samples, sobolnboot, sobolconf, nseeds, precision)
}
\arguments{
\item{nl}{nl object with a defined experiment}

\item{samples}{number of samples for the sobol sensitivity analysis}

\item{sobolnboot}{number of bootstrap replicates of the sobol sensitivity analysis}

\item{sobolconf}{the confidence level for bootstrap confidence intervals}

\item{nseeds}{number of seeds for this simulation design}

\item{precision}{number of digits for the decimal fraction of parameter values}
}
\value{
simdesign S4 class object
}
\description{
Add a soboljansen simdesign to a nl object
}
\details{
This function creates a simdesign S4 class which can be added to a nl object by using the setter function simdesign(nl).
The soboljansen simdesign uses parameter ranges and q-functions that are defined in the variables field of the experiment object within the nl object.
It uses the sensitivity package to set up a soboljansen sensitivity analysis, including a simobject of class sobol and a input tibble for simulations.
For details on method specific sensitivity analysis function parameters see ?soboljansen
Finally, the function reports a simdesign object.
}
\examples{
\dontrun{
# Example for Wolf Sheep Predation model from NetLogo models library:
nl@simdesign <- simdesign_soboljansen(nl=nl,
samples=200,
sobolnboot=20,
sobolconf=0.95,
nseeds=3,
precision=3)
}

}
