% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_netlogo.R
\name{supported_netlogo_versions}
\alias{supported_netlogo_versions}
\title{Report supported NetLogo versions}
\usage{
supported_netlogo_versions()
}
\value{
a vector of supported Versions for Download and Usage (parameter \code{version}):
}
\description{
Auxiliary function to report supported NetLogo version numbers
}
\examples{
\dontrun{
supported_netlogo_versions()
}

}
