% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_nlrx.R
\name{test_nlrx}
\alias{test_nlrx}
\title{Test if nlrx runs on the local system}
\usage{
test_nlrx(nlpath, nlversion)
}
\arguments{
\item{nlpath}{Provide a path to a netlogo folder}

\item{nlversion}{Matching version string of the provided NetLogo folder (e.g. "7.0.4")}
}
\description{
Runs a short test if nlrx runs on the local system
}
\details{
Runs a short test if nlrx runs on the local system. Reports TRUE if successful!
}
\examples{
\dontrun{
test_nlrx(nlpath="/Users/xyz/netlogo/NetLogo 7.0.4", nlversion="7.0.4")
}

}
