% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_simoutput.R
\name{unnest_simoutput}
\alias{unnest_simoutput}
\title{Get spatial data from metrics.turtles and metrics.patches output}
\usage{
unnest_simoutput(nl)
}
\arguments{
\item{nl}{nl object}
}
\value{
tibble with spatial data objects
}
\description{
Turn results from NetLogo in spatial data objects
}
\details{
Unnests output from run_nl into long format.
}
\examples{

}
