% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl.R
\name{util_call_nl}
\alias{util_call_nl}
\title{Setup and execute NetLogo via commandline}
\usage{
util_call_nl(nl, xmlfile, outfile, batchfile)
}
\arguments{
\item{nl}{nl object}

\item{xmlfile}{file location of the experiment xml file}

\item{outfile}{file location for output results}

\item{batchfile}{file location of system specific batch file to call NetLogo via commandline}
}
\description{
Setup and execute NetLogo via commandline
}
\keyword{internal}
