% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stuff.R
\name{util_check_deprecated_args}
\alias{util_check_deprecated_args}
\title{Utility function for checking for deprecated or unsupported arguments}
\usage{
util_check_deprecated_args(dots, deprecated_args)
}
\arguments{
\item{dots}{Named list of arguments captured from \code{...}}

\item{deprecated_args}{Character vector of deprecated argument names that should trigger depreciation warnings (and not errors)}
}
\description{
Utility function for checking for deprecated or unsupported arguments
}
\keyword{internal}
