% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl.R
\name{util_clean_agent_metrics}
\alias{util_clean_agent_metrics}
\title{Backend function for cleaning agent metrics}
\usage{
util_clean_agent_metrics(NLtable, nl, expect_cleaned_names = FALSE)
}
\arguments{
\item{NLtable}{simulation output table}

\item{nl}{nl object}

\item{expect_cleaned_names}{TRUE/FALSE, if TRUE reporter column names are expected to be cleaned by \code{janitor::make_clean_names()}.}
}
\value{
simulation output table with cleaned agent metrics
}
\description{
Internal helper for formatting turtle, patch and link metrics in simulation output.
}
\details{
Converts raw turtle, patch and link metric reporter output into nested output columns.
}
\keyword{internal}
