% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl.R
\name{util_collect_experiment_metrics}
\alias{util_collect_experiment_metrics}
\title{Backend function for collecting experiment metrics}
\usage{
util_collect_experiment_metrics(nl)
}
\arguments{
\item{nl}{nl object}
}
\value{
character vector of NetLogo reporter metrics
}
\description{
Internal helper for combining regular metrics with turtle, patch and link metrics.
}
\details{
Combines regular experiment metrics with additional agent metrics defined in
\code{metrics.turtles}, \code{metrics.patches} and \code{metrics.links}.
}
\keyword{internal}
