% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl.R
\name{util_create_agentset_reporter}
\alias{util_create_agentset_reporter}
\title{Create turtle/patches/links owns reporter}
\usage{
util_create_agentset_reporter(metrics, breed)
}
\arguments{
\item{metrics}{character vector, the names of variables to measure}

\item{breed}{character, the corresponding breed/agentset (e.g. "turtles", "patches", "links", "wolves", ...)}
}
\description{
The reporter is used to measure agent variables in NetLogo and parse them to a readable format
}
\keyword{internal}
