% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stuff.R
\name{util_create_lhs}
\alias{util_create_lhs}
\title{Identify and report the current OS}
\usage{
util_create_lhs(input, samples, precision)
}
\arguments{
\item{input}{list with variables and value ranges}

\item{samples}{number of lhs samples}

\item{precision}{number of digits for the decimal fraction of parameter
values}
}
\description{
Identify and report the current OS
}
\keyword{internal}
