% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl.R
\name{util_create_sim_XML}
\alias{util_create_sim_XML}
\title{Create a temporary behavior space xml file to setup NetLogo via command line}
\usage{
util_create_sim_XML(nl, seed, siminputrow, xmlfile)
}
\arguments{
\item{nl}{nl object}

\item{seed}{random-seed for NetLogo simulation}

\item{siminputrow}{row id of the simulation input tibble of the simdesign
within the provided nl object}

\item{xmlfile}{filepath where the xml file is stored}
}
\description{
Create a temporary behavior space xml file to setup NetLogo via
command line
}
\keyword{internal}
