% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_eval.R
\name{util_eval_simdesign}
\alias{util_eval_simdesign}
\title{Evaluate all slots of a simdesign object}
\usage{
util_eval_simdesign(nl)
}
\arguments{
\item{nl}{nl object}
}
\description{
Evaluate all slots of a simdesign object
}
\details{
util_eval_simdesign checks if the information stored within the simdesign slots are valid.
}
\keyword{internal}
