% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_eval.R
\name{util_eval_variables}
\alias{util_eval_variables}
\title{Evaluate if variables list of an experiment object is empty}
\usage{
util_eval_variables(nl)
}
\arguments{
\item{nl}{nl object}
}
\description{
Evaluate if variables list of an experiment object is empty
}
\details{
util_eval_variables checks if the variables list of an experiment within a nl object is empty.
It reports an error message if no variables were defined.
}
\keyword{internal}
