% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_eval.R
\name{util_eval_variables_sa}
\alias{util_eval_variables_sa}
\title{Evaluate variables list of an experiment object}
\usage{
util_eval_variables_sa(nl)
}
\arguments{
\item{nl}{nl object}
}
\description{
Evaluate variables list of an experiment object
}
\keyword{internal}
