% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl.R
\name{util_gather_results}
\alias{util_gather_results}
\title{Load output file from simulations}
\usage{
util_gather_results(nl, outfile)
}
\arguments{
\item{nl}{nl object}

\item{outfile}{file location of output results}
}
\description{
Load output file from simulations
}
