% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stuff.R
\name{util_generate_seeds}
\alias{util_generate_seeds}
\title{Generate a vector of random seeds}
\usage{
util_generate_seeds(nseeds)
}
\arguments{
\item{nseeds}{desired length of the random seeds vector}
}
\description{
Generate a vector of random seeds
}
\keyword{internal}
