% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_nl.R
\name{util_print.experiment}
\alias{util_print.experiment}
\title{Print experiment object content}
\usage{
util_print.experiment(x, ...)
}
\arguments{
\item{x}{experiment object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print experiment object content
}
\details{
Print content of experiment object in a nice formatted overview to console
}
\keyword{internal}
