% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_nl.R
\name{util_print.nl}
\alias{util_print.nl}
\title{Print nl object content}
\usage{
util_print.nl(x, ...)
}
\arguments{
\item{x}{nl object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print nl object content
}
\details{
Print content of nl object in a nice formatted overview to console
}
\keyword{internal}
