% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_r.R
\name{util_print.r}
\alias{util_print.r}
\title{Print r object content}
\usage{
util_print.r(x, ...)
}
\arguments{
\item{x}{r object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print r object content
}
\details{
Print content of r object in a nice formatted overview to console
}
\keyword{internal}
