% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_nl.R
\name{util_print.simdesign}
\alias{util_print.simdesign}
\title{Print simdesign object content}
\usage{
util_print.simdesign(x, ...)
}
\arguments{
\item{x}{simdesign object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print simdesign object content
}
\details{
Print content of simdesign object in a nice formatted overview to console
}
\keyword{internal}
