% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_nl.R, R/print_r.R
\name{util_print.summary}
\alias{util_print.summary}
\title{Print nl object summary}
\usage{
util_print.summary(x, ...)

util_print.summary(x, ...)
}
\arguments{
\item{x}{r object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print nl object summary

Print r object summary
}
\details{
Print summary of nl object and embedded experiment and simdesign objects in a nice formatted overview to console

Print summary of r object and embedded experiment and simdesign objects in a nice formatted overview to console
}
\keyword{internal}
