% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl.R
\name{util_read_write_batch}
\alias{util_read_write_batch}
\title{Write a modified batchfile that executes NetLogo}
\usage{
util_read_write_batch(nl)
}
\arguments{
\item{nl}{nl object}
}
\description{
Write a modified batchfile that executes NetLogo
}
\keyword{internal}
