% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl_dyn.R
\name{util_run_nl_dyn_GenAlg_fn}
\alias{util_run_nl_dyn_GenAlg_fn}
\alias{util_run_nl_dyn_ABCmcmc_fn}
\title{Genetic Algorithm run simulation function}
\usage{
util_run_nl_dyn_GenAlg_fn(param, nl, evalcrit, seed, threads = 1, ...)

util_run_nl_dyn_ABCmcmc_fn(param, threads = 1)
}
\arguments{
\item{param}{vector of model parameters passed from ABC_mcmc function. If use_seeds = TRUE, the first element of this vector is a random seed}

\item{nl}{nl object}

\item{evalcrit}{evaluation criterion for simulated annealing}

\item{seed}{vector of model seeds, one per replicated run of this evaluation}

\item{threads}{number of NetLogo threads used for execution (handled via NetLogo).}
}
\description{
Genetic Algorithm run simulation function

Genetic Algorithm run simulation function
}
\keyword{internal}
