% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{write_simoutput}
\alias{write_simoutput}
\title{Write NetLogo simulation output to file}
\usage{
write_simoutput(nl)
}
\arguments{
\item{nl}{nl object

Write NetLogo simulation output to a csv file in the direcotry outpath of the nl object
Output has to be attached to the simdesign first with simoutput(nl) <- results}
}
\description{
Write NetLogo simulation output to file
}
\examples{
\dontrun{

# Run one simulation:
results <- run_nl(
  nl = nl,
  seed = getsim(nl, "simseeds")[1],
  run = 1,
  cleanup = "all"
)

# Attach output to simdesign:
setsim(nl, "simoutput") <- results

# Write output to outpath directory
write_simoutput(nl)
}
}
