% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{analyze_eFast}
\alias{analyze_eFast}
\title{Analyze NetLogo simulation output of simdesign eFast}
\usage{
analyze_eFast(nl, metrics, funs)
}
\arguments{
\item{nl}{nl object}

\item{metrics}{vector of strings defining metric columns for evaluation. Defaults to metrics of the experiment within the nl object}

\item{funs}{dplyr::funs list with the summary metrics one wants to have for the sensitivity results}
}
\description{
Analyze NetLogo simulation output of simdesign eFast
}
\details{
The function calculates eFast sensitivity indices from the output results using the \link[sensitivity]{sensitivity} package.
}
\keyword{internal}
