% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{analyze_ff}
\alias{analyze_ff}
\title{Analyze NetLogo simulation output of simdesign full-factorial}
\usage{
analyze_ff(nl, metrics, funs)
}
\arguments{
\item{nl}{nl object}

\item{metrics}{vector of strings defining metric columns for evaluation. Defaults to metrics of the experiment within the nl object}

\item{funs}{list with the summary metrics for the sensitivity results}
}
\description{
Analyze NetLogo simulation output of simdesign full-factorial
}
\details{
The function calculates aggregated output metrics by dropping random seeds and aggregating values with the provided functions.
}
\keyword{internal}
