% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{analyze_ff}
\alias{analyze_ff}
\title{Analyze NetLogo simulation output of simdesign full-factorial}
\usage{
analyze_ff(nl)
}
\arguments{
\item{nl}{nl object}
}
\description{
Analyze NetLogo simulation output of simdesign full-factorial
}
\keyword{internal}
