% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{analyze_lhs}
\alias{analyze_lhs}
\title{Analyze NetLogo simulation output of simdesign latin-hypercube}
\usage{
analyze_lhs(nl)
}
\arguments{
\item{nl}{nl object}
}
\description{
Analyze NetLogo simulation output of simdesign latin-hypercube
}
