% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{analyze_nl}
\alias{analyze_nl}
\title{Analyze NetLogo simulation output}
\usage{
analyze_nl(nl)
}
\arguments{
\item{nl}{nl object

Runs basic analyses on NetLogo simulation output
Output has to be attached to the simdesign first with simoutput(nl) <- results

The functions calls post-processing analysis functions, depending on the specified method in the simdesign object of the nl object.}
}
\description{
Analyze NetLogo simulation output
}
\examples{
\dontrun{

# Run one simulation:
results <- run_nl(nl=nl,
seed=simseeds(nl)[1],
run=1,
cleanup="all")

# Attach output to simdesign:
simoutput(nl) <- results

# Perform analysis:
analyze_nl(nl)

}
}
