% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{analyze_nl}
\alias{analyze_nl}
\title{Analyze NetLogo simulation output}
\usage{
analyze_nl(nl, metrics = getexp(nl, "metrics"),
  funs = dplyr::funs(mean))
}
\arguments{
\item{nl}{nl object}

\item{metrics}{vector of strings defining metric columns for evaluation. Defaults to metrics of the experiment within the nl object}

\item{funs}{dplyr::funs list with the summary metrics one wants to have for the sensitivity results}
}
\value{
analysis summary tibble

Runs basic analyses on NetLogo simulation output
Output has to be attached to the simdesign first with \code{setsim(nl, "output") <- results}

The functions calls post-processing analysis functions, depending on the specified method in the simdesign object of the nl object.
}
\description{
Analyze NetLogo simulation output
}
\examples{
\dontrun{

# Run one simulation:
results <- run_nl(
  nl = nl,
  seed = getsim(nl, "simseeds")[1],
  run = 1,
  cleanup = "all"
)

# Attach output to simdesign:
setsim(nl, "simoutput") <- results

# Perform analysis:
myfuns <- dplyr::funs(mean, sd, min, max)
analyze_nl(nl, myfuns)
}
}
