% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{analyze_simple}
\alias{analyze_simple}
\title{Analyze NetLogo simulation output of simdesign simple}
\usage{
analyze_simple(nl)
}
\arguments{
\item{nl}{nl object}
}
\description{
Analyze NetLogo simulation output of simdesign simple
}
