% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{analyze_sobol}
\alias{analyze_sobol}
\title{Analyze NetLogo simulation output of simdesign sobol}
\usage{
analyze_sobol(nl)
}
\arguments{
\item{nl}{nl object}
}
\description{
Analyze NetLogo simulation output of simdesign sobol
}
\keyword{internal}
