% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{analyze_sobol2007}
\alias{analyze_sobol2007}
\title{Analyze NetLogo simulation output of simdesign sobol2007}
\usage{
analyze_sobol2007(nl)
}
\arguments{
\item{nl}{nl object}
}
\description{
Analyze NetLogo simulation output of simdesign sobol2007
}
\keyword{internal}
