% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{analyze_soboljansen}
\alias{analyze_soboljansen}
\title{Analyze NetLogo simulation output of simdesign soboljansen}
\usage{
analyze_soboljansen(nl)
}
\arguments{
\item{nl}{nl object}
}
\description{
Analyze NetLogo simulation output of simdesign soboljansen
}
\keyword{internal}
