% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_netlogo.R
\name{check_netlogo_version}
\alias{check_netlogo_version}
\title{Check supported NetLogo versions}
\usage{
check_netlogo_version(version, throw_error = FALSE)
}
\arguments{
\item{version}{Character string naming which NetLogo Version to check}

\item{throw_error}{TRUE/FALSE, if TRUE an error is thrown if the version is not supported}
}
\value{
logical, if netlogo versions is supported by nlrx
}
\description{
Auxiliary function to check supported NetLogo version numbers
}
\examples{
\dontrun{
check_netlogo_version("6.2.0")
check_netlogo_version("1.0")
check_netlogo_version("1.0", TRUE)
}

}
