% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_netlogo.R
\name{download_netlogo}
\alias{download_netlogo}
\title{Download NetLogo}
\usage{
download_netlogo(to, version)
}
\arguments{
\item{to}{Character string with the name where the downloaded file is saved.}

\item{version}{Character string naming which NetLogo Version to download (see Details)}
}
\description{
Auxiliary function to download NetLogo
}
\details{
\itemize{
\item "604" = NetLogo Version 6.0.4
\item "603" = NetLogo Version 6.0.3
\item "602" = NetLogo Version 6.0.2
\item "601" = NetLogo Version 6.0.1
\item "600" = NetLogo Version 6.0.0
\item "531" = NetLogo Version 5.3.1
}
}
\examples{
\dontrun{
download_netlogo("/home/user/experiment/" , "603")
}

}
