% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_netlogo.R
\name{download_netlogo}
\alias{download_netlogo}
\title{Download NetLogo}
\usage{
download_netlogo(to, version, extract = FALSE)
}
\arguments{
\item{to}{Path to folder where the downloaded file is saved.}

\item{version}{Character string naming which NetLogo Version to download (see Details)}

\item{extract}{TRUE/FALSE, if TRUE downloaded archive is extracted to subfolder of \code{to} (only unix)}
}
\description{
Auxiliary function to download NetLogo
}
\details{
Retrieve supported Versions for Download and Usage (parameter \code{version}):
supported_netlogo_versions()
}
\examples{
\dontrun{
dlpath <- tempdir()  # adjust path to your needs
try(download_netlogo(dlpath, "6.0.3"))
}

}
