% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constr.R
\name{experiment}
\alias{experiment}
\title{Construct a new experiment object}
\usage{
experiment(expname = "defaultexp", outpath = NA_character_,
  repetition = 1, tickmetrics = "true", idsetup = "setup",
  idgo = "go", idfinal = NA_character_, idrunnum = NA_character_,
  runtime = 1, evalticks = seq(1, runtime, 1),
  stopcond = NA_character_, metrics = c("count turtles"),
  metrics.turtles = NA_character_, metrics.patches = NA_character_,
  metrics.links = NA_character_, variables = list(),
  constants = list(), ...)
}
\arguments{
\item{expname}{A character string defining the name of the experiment}

\item{outpath}{Path to a directory where experiment output will be stored}

\item{repetition}{A number which gives the number of repetitions for each row of the simulation design input tibble}

\item{tickmetrics}{Character string "true" runs defined metrics on each simulation tick. "false" runs metrics only after simulation is finished}

\item{idsetup}{character string or vector of character strings, defining the name of the NetLogo setup procedure}

\item{idgo}{character string or vector of character strings, defining the name of the NetLogo go procedure}

\item{idfinal}{character string or vector of character strings, defining the name of NetLogo procedures that should be run after the last tick}

\item{idrunnum}{character string, defining the name of a NetLogo global that should be used to parse the current siminputrow during model executions which can then be used for self-written output.}

\item{runtime}{number of model ticks that should be run for each simulation}

\item{evalticks}{vector of tick numbers defining when measurements are taken}

\item{stopcond}{a NetLogo reporter that reports TRUE/FALSE. If it reports TRUE the current simulation run is stopped (e.g. "not any? turtles")}

\item{metrics}{vector of strings defining valid NetLogo reporters that are taken as output measurements (e.g. c("count turtles", "count patches"))}

\item{metrics.turtles}{vector of strings defining valid turtles-own variables that are taken as output measurements (e.g. c("who", "pxcor", "pycor", "color"))}

\item{metrics.patches}{vector of strings defining valid patches-own variables that are taken as output measurements (e.g. c("pxcor", "pycor", "pcolor"))}

\item{metrics.links}{vector of strings defining valid links-own variables that are taken as output measurements (e.g. c("end1", "end2"))}

\item{variables}{a nested list of variables that are changed within a simulation design. The name of each sublist item has to be a valid global of the defined NetLogo model. Depending on the desired simdesign each list item consist of a vector of values, a min value, a max value, a step value and a qfun (e.g. list("paramA" = list(values=c(0, 0.5, 1), min=0, max=1, step=0.1, qfun="qunif")))}

\item{constants}{a list of constants that are kept constant within a simulation design. The name of each list item has to be a valid global of the defined NetLogo model (e.g. list("pNUM" = 12, "pLOGIC"="TRUE", "pSTRING"=""default""))}

\item{...}{...}
}
\value{
experiment S4 class object
}
\description{
Construct a new experiment object
}
\details{
The experiment class stores all information related to the NetLogo simulation experiment.
The class holds all information that is typically entered into NetLogo Behavior Space experiments.
When setting up an experiment, it is usually attached to an already defined \link[nlrx]{nl} object (see examples).
After attaching an experiment, different simdesign helper functions can be used to attach a simdesign to the nl object \link[nlrx]{simdesign}.
The simdesign helper functions use the variable definitions from the experiment within the nl object to generate a parameter tibble for simulations.

\strong{The following class slots are obligatory to run an experiment:}

\emph{repetition}

In cases, where the random seed is controlled by nlrx simdesigns, repitition should be set to one as random seeds would not differ between simulations.
In cases, where the random seed is set within the NetLogo model, repitition can be increased to repeat the same parameterisation with different random seeds.

\emph{tickmetrics}

If "true", the defined output reporters are collected on each simulation tick that is defined in evalticks. If "false" measurements are taken only on the last tick.

\emph{idsetup, idgo}

These two class slots accept strings, or vectors of strings, defining NetLogo model procedures that should be executed for model setup (idestup) and model execution (idgo).

\emph{runtime}

Defines the maximum number of simulation ticks that are executed.

\strong{Depending on the simdesign, the following slots may be obligatory:}

\emph{metrics}

A vector of valid netlogo reporters that defines which measurements are taken.
The optimization simdesigns need at least one defined metrics reporter for fitness calculation of the optimization algorithm.

\emph{constants, variables}

These slots accept lists with NetLogo parameters that should be varied within a simdesign (variables) or should be kept constant (constants) for each simulation.
Any NetLogo parameter that is not entered in at least one of these two lists will be set up as constant with the default value from the NetLogo interface.
It is not possible to enter a NetLogo parameter in both lists (a warning message will appear when a simdesign is attached to such an experiment).
All simdesigns except \link[nlrx]{simdesign_simple} need defined variables for setting up a parameter matrix.
Variables can be defined as distinct values, value distributions or range with increment.
The information that is needed, depends on the chosen simdesign (details on variable definition requiements can be found in the helpfiles of each simdesign helper function).

\strong{All remaining slots are optional:}

\emph{expname}

A character string defining the name of the experiment, useful for documentation purposes.

\emph{outpath}

A valid path to an existing directory. The directory is used by the \link[nlrx]{write_simoutput} function to store attached simulation results to disk in csv format.

\emph{idfinal}

A character string or vector of strings defining NetLogo procedures that are executed at the end of each simulation (e.g. cleanup or self-written output procedures).

\emph{idrunnum}

This slot can be used to transfer the current nlrx experiment name, random seed and runnumber (siminputrow) to NetLogo.
To use this functionality, a string input field widget needs to be created on the GUI of your NetLogo model.
The name of this widget can be entered into the "idrunnum" field of the experiment.
During simulations, the value of this widget is automatically updated with a generated string that contains the current nlrx experiment name, random seed and siminputrow ("expname_seed_siminputrow").
For self-written output In NetLogo, we suggest to include this global variable which allows referencing the self-written output files to the collected output of the nlrx simulations in R.

\emph{evalticks}

A vector of integers, defining the ticks for which the defined metrics will be measured.

\emph{stopcond}

The stopcond slot can be used to define a stop condition by providing a string with valid NetLogo code that reports either true or false.
Each simulation will be stopped automatically, once the reporter reports true.

\emph{metrics.turtles, metrics.patches}

These two slots can be used to enter turtles-own (metrics.turtles) and patches-own (metrics.patches) variables of the NetLogo model.
These agent variables are measured in addition to the defined metrics.
After attaching the simulation results to the nl object, the measured output from these agent variables needs to be postprocessed by \link[nlrx]{get_nl_spatial}.
Please note that NetLogo models may contain a huge number of patches and turtles and output measurements of agent variables on each tick may need a lot of ressources.
}
\examples{
\dontrun{
# Example for Wolf Sheep Predation model from NetLogo models library:
nl@experiment <- experiment(expname="wolf-sheep",
                             outpath="C:/out/",
                             repetition=1,
                             tickmetrics="true",
                             idsetup="setup",
                             idgo="go",
                             idfinal=NA_character_,
                             idrunnum=NA_character_,
                             runtime=50,
                             evalticks=seq(40,50),
                             stopcond="not any? turtles",
                             metrics=c("count sheep",
                                       "count wolves",
                                     "count patches with [pcolor = green]"),
                             metrics.turtles=c("who",
                                               "pxcor",
                                               "pycor",
                                               "color"),
                             metrics.patches=c("pxcor", "pycor", "pcolor"),
                             metrics.links=c("end1","end2"),
                             variables = list('initial-number-sheep' =
                             list(min=50, max=150, step=10, qfun="qunif"),
                                              'initial-number-wolves' =
                             list(min=50, max=150, step=10, qfun="qunif")),
                             constants = list("model-version" =
                                              "\\"sheep-wolves-grass\\"",
                                              "grass-regrowth-time" = 30,
                                              "sheep-gain-from-food" = 4,
                                              "wolf-gain-from-food" = 20,
                                              "sheep-reproduce" = 4,
                                              "wolf-reproduce" = 5,
                                              "show-energy?" = "false"))

}

}
