% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_setget.R
\name{getexp}
\alias{getexp}
\title{Getter function to get a variable of an experiment object}
\usage{
getexp(nl, var)
}
\arguments{
\item{nl}{nl object}

\item{var}{valid experiment variable string}
}
\description{
Getter function to get a variable of an experiment object
}
\examples{
\dontrun{
getexp(nl, "expname")
}
}
