% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_setget.R
\name{getexp}
\alias{getexp}
\title{Getter function to get a variable of an experiment object}
\usage{
getexp(nl, var)
}
\arguments{
\item{nl}{nl object}

\item{var}{valid experiment variable string}
}
\description{
Getter function to get a variable of an experiment object
}
\examples{
# Example for Wolf Sheep Predation model from NetLogo models library:
nl <- nl(nlversion = "6.0.3",
nlpath = "/home/user/NetLogo 6.0.3/",
modelpath = "/home/user/NetLogo 6.0.3/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo",
jvmmem = 1024)

# Set experiment name
setexp(nl, "expname") <- "experimentName"

# Get experiment name
getexp(nl, "experiment")

}
