% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_setget.R
\name{getnl}
\alias{getnl}
\title{Getter function to get a variable of a nl object}
\usage{
getnl(nl, var)
}
\arguments{
\item{nl}{nl object}

\item{var}{valid nl variable string}
}
\description{
Getter function to get a variable of a nl object
}
\examples{
\dontrun{
getnl(nl, "nlversion")
}
}
