% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_setget.R
\name{getnl}
\alias{getnl}
\title{Getter function to get a variable of a nl object}
\usage{
getnl(nl, var)
}
\arguments{
\item{nl}{nl object}

\item{var}{valid nl variable string}
}
\description{
Getter function to get a variable of a nl object
}
\examples{

# Example for Wolf Sheep Predation model from NetLogo models library:
nl <- nl(nlversion = "6.0.3",
nlpath = "/home/user/NetLogo 6.0.3/",
modelpath = "/home/user/NetLogo 6.0.3/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo",
jvmmem = 1024)

# get NetLogo version
getnl(nl, "nlversion")

}
