% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_setget.R
\name{getsim}
\alias{getsim}
\title{Getter function to get a variable of a simdesign object}
\usage{
getsim(nl, var)
}
\arguments{
\item{nl}{nl object}

\item{var}{valid simdesign variable string}
}
\description{
Getter function to get a variable of a simdesign object
}
\examples{
# Example for Wolf Sheep Predation model from NetLogo models library:
nl <- nl(nlversion = "6.0.3",
nlpath = "/home/user/NetLogo 6.0.3/",
modelpath = "/home/user/NetLogo 6.0.3/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo",
jvmmem = 1024)

# Set simulation seeds
setsim(nl, "simseeds") <- c(123, 456, 789)

# Set simulation seeds
getsim(nl, "simseeds")

}
