% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_setget.R
\name{getsim}
\alias{getsim}
\title{Getter function to get a variable of a simdesign object}
\usage{
getsim(nl, var)
}
\arguments{
\item{nl}{nl object}

\item{var}{valid simdesign variable string}
}
\description{
Getter function to get a variable of a simdesign object
}
\examples{
\dontrun{
getsim(nl, "simseeds")
}
}
