% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constr.R
\name{nl}
\alias{nl}
\title{Construct a new nl object}
\usage{
nl(nlversion = "6.0.2", nlpath = character(), modelpath = character(),
  jvmmem = 1024, ...)
}
\arguments{
\item{nlversion}{A character string defining the NetLogo version that is used}

\item{nlpath}{Path to the NetLogo main directory matching the defined version}

\item{modelpath}{Path to the NetLogo model file (*.nlogo) that is used for simulations}

\item{jvmmem}{Java virtual machine memory capacity in megabytes}

\item{experiment}{Holds a experiment S4 class object}
}
\description{
Construct a new nl object
}
\details{
nl objects are the main class objects used in the nlrx package.
These objects store all information that is needed to run NetLogo simulations.
nl objects are initialized with basic information on Netlogo and the model.
To run simulation experiments, a experiment needs to be added to the nl object.
Each experiment also needs to have a specified simulation design in order to run simulations.
}
\examples{
\dontrun{
# Example for Wolf Sheep Predation model from NetLogo models library:
nl <- nl(nlversion = "6.0.3",
nlpath = "C:/Program Files/NetLogo 6.0.3/",
modelpath = "C:/Program Files/NetLogo 6.0.3/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo",
jvmmem = 1024)
}

}
