% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlrx-package.R
\docType{package}
\name{nlrx-package}
\alias{nlrx-package}
\title{nlrx: A package for running NetLogo simulations from R.}
\description{
The nlrx package provides tools to setup NetLogo simulations in R.
It uses a similar structure as NetLogos Behavior Space but offers more flexibility and additional tools for running sensitivity analyses.
}
\author{
Jan Salecker \email{jsaleck@gwdg.de}
}
\keyword{package}
