% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constr.R
\name{r}
\alias{r}
\title{Construct a new r object
#'}
\usage{
r(
  nlversion = "6.1.1",
  nlpath = character(),
  modelpath = character(),
  obj_type = "r_obj",
  ...
)
}
\arguments{
\item{nlversion}{A character string defining the NetLogo version that is used}

\item{nlpath}{Path to the NetLogo main directory matching the defined version}

\item{modelpath}{Path to the NetLogo model file (*.nlogo) that is used for simulations}

\item{obj_type}{A character string defining the object (nl, R, ...) (YYY)}

\item{...}{...}

\item{jvmmem}{Java virtual machine memory capacity in megabytes}

\item{experiment}{Holds a experiment S4 class object}

\item{simdesign}{Holds a simdesign S4 class object}
}
\value{
r S4 class object
}
\description{
Construct a new r object
}
\details{
r objects...
}
\examples{
# Example for Wolf Sheep Predation model from NetLogo models library:
# Windows default NetLogo installation path (adjust to your needs!):
netlogopath <- file.path("C:/Program Files/NetLogo 6.0.3")
modelpath <- file.path(netlogopath, "app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo")
outpath <- file.path("C:/out")
# Unix default NetLogo installation path (adjust to your needs!):
netlogopath <- file.path("/home/NetLogo 6.0.3")
modelpath <- file.path(netlogopath, "app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo")
outpath <- file.path("/home/out")


r <- r(nlversion = "6.1.1",
         nlpath = netlogopath,
         modelpath = modelpath,
         obj_type = "r_obj",
         jvmmem = 1024)

}
