% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_all}
\alias{run_all}
\title{Execute simulations depending on the object}
\usage{
run_all(obj)
}
\arguments{
\item{obj}{nl or R object}
}
\value{
tibble with simulation output results
}
\description{
Execute simulations depending on the object (nl or R object)
}
