% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nl.R
\name{run_nl_all}
\alias{run_nl_all}
\title{Execute NetLogo simulation}
\usage{
run_nl_all(nl, cleanup = "all")
}
\arguments{
\item{nl}{nl object}

\item{cleanup}{indicate which filetypes should be deleted}
}
\value{
tibble with simulation output results
}
\description{
Execute NetLogo simulation from a nl object with a defined experiment and simdesign
}
\details{
run_nl_all executes all simulations of the specified NetLogo model within the provided nl object.
The function loops over all random seeds and all rows of the siminput table of the simdesign of nl.
The loops are created by calling furr::future_map_dfr which allows running the function either locally or on remote HPC machines.
Cleanup can either be ".xml" to delete all temporarily created xml files; ".csv" to delete all temporarily created csv files or "all" to delete all temporarily created files.
}
\examples{
\dontrun{

# Run parallel on local machine:
future::plan(multisession)
results \%<-\% run_nl_all(nl, cleanup="all")

}
}
