% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nl.R
\name{run_nl_dyn}
\alias{run_nl_dyn}
\title{Execute NetLogo simulation without pregenerated parametersets}
\usage{
run_nl_dyn(nl, seed, cleanup = "all")
}
\arguments{
\item{nl}{nl object}

\item{seed}{a random seed for the NetLogo simulation}

\item{cleanup}{indicate which filetypes should be deleted}
}
\value{
simulation output results can be tibble, list, ...
}
\description{
Execute NetLogo simulation from a nl object with a defined experiment and simdesign but no pregenerated input parametersets
}
\details{
run_nl_dyn can be used for simdesigns where no predefined parametersets exist.
This is the case for dynamic designs, such as SImulated Annealing, where parametersets are dynamically generated, based on the output of previous simulations.
Cleanup can either be ".xml" to delete all temporarily created xml files; ".csv" to delete all temporarily created csv files or "all" to delete all temporarily created files.
}
\examples{
\dontrun{

# Run one simulation:
results <- run_nl_dyn(
  nl = nl,
  cleanup = "all"
)
}
}
