% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nl.R
\name{run_nl_one}
\alias{run_nl_one}
\title{Execute one NetLogo simulation from a nl object}
\usage{
run_nl_one(nl, seed, siminputrow, cleanup = "all")
}
\arguments{
\item{nl}{nl object}

\item{seed}{a random seed for the NetLogo simulation}

\item{siminputrow}{rownumber of the input tibble within the attached simdesign object that should be executed}

\item{cleanup}{indicate which filetypes should be deleted}
}
\value{
tibble with simulation output results
}
\description{
Execute one NetLogo simulation from a nl object with a defined experiment and simdesign
}
\details{
run_nl_one executes one simulation of the specified NetLogo model within the provided nl object.
The random seed is set within the NetLogo model to control stochasticity.
The siminputrow number defines which row of the input data tibble within the simdesign object of the provided nl object is executed.
Cleanup can either be "xml" to delete all temporarily created xml files; "csv" to delete all temporarily created csv files,
"bat" to create all temporarily created bat/sh files or "all" to delete all temporarily created files.

This function can be used to run single simulations of a NetLogo model.
}
\examples{
\dontrun{

# Run one simulation:
results <- run_nl_one(
  nl = nl,
  seed = getsims(nl, "simseeds")[1],
  siminputrow = 1,
  cleanup = "all"
)
}
}
