% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_r.R
\name{run_r_one}
\alias{run_r_one}
\title{Execute one r simulation from a r object}
\usage{
run_r_one(
  r,
  seed,
  siminputrow,
  cleanup.csv = TRUE,
  cleanup.xml = TRUE,
  cleanup.bat = TRUE,
  writeRDS = FALSE
)
}
\arguments{
\item{r}{nl object}

\item{seed}{a random seed for the NetLogo simulation}

\item{siminputrow}{rownumber of the input tibble within the attached simdesign object that should be executed}

\item{cleanup.csv}{TRUE/FALSE, if TRUE temporary created csv output files will be deleted after gathering results.}

\item{cleanup.xml}{TRUE/FALSE, if TRUE temporary created xml output files will be deleted after gathering results.}

\item{cleanup.bat}{TRUE/FALSE, if TRUE temporary created bat/sh output files will be deleted after gathering results.}

\item{writeRDS}{TRUE/FALSE, if TRUE an rds file with the simulation results will be written to the defined outpath folder of the experiment within the nl object.}
}
\value{
tibble with simulation output results
}
\description{
Execute one r simulation from a r object with a defined experiment and simdesign
}
\details{
run_r_one executes one simulation...
}
\examples{
\dontrun{

# Load nl object from test data:
r <- nl_lhs

# Run one simulation:
results <- run_r_one(r = r,
                      seed = getsim(r, "simseeds")[1],
                      siminputrow = 1)

}
}
