% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_setget.R
\name{setexp<-}
\alias{setexp<-}
\alias{setexp}
\title{Setter function to set a variable of an experiment object}
\usage{
setexp(nl, var) <- value
}
\arguments{
\item{nl}{nl object}

\item{var}{valid experiment variable string}

\item{value}{valid value for the specified variable}
}
\description{
Setter function to set a variable of an experiment object
}
\examples{
\dontrun{
setexp(nl, "expname") <- "experimentName"
}
}
