% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_setget.R
\name{setnl<-}
\alias{setnl<-}
\alias{setnl}
\title{Setter function to set a variable of a nl object}
\usage{
setnl(nl, var) <- value
}
\arguments{
\item{nl}{nl object}

\item{var}{valid nl variable string}

\item{value}{valid value for the specified variable}
}
\description{
Setter function to set a variable of a nl object
}
\examples{
# Example for Wolf Sheep Predation model from NetLogo models library:
nl <- nl(
nlpath = "/home/user/NetLogo 6.0.3/",
modelpath = "/home/user/NetLogo 6.0.3/app/models/Sample Models/Biology/Wolf Sheep Predation.nlogo",
jvmmem = 1024)

# set NetLogo version
setnl(nl, "nlversion") <- "6.0.3"

}
