% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_setget.R
\name{setsim<-}
\alias{setsim<-}
\alias{setsim}
\title{Setter function to set a variable of a simdesign object}
\usage{
setsim(nl, var) <- value
}
\arguments{
\item{nl}{nl object}

\item{var}{valid simdesign variable string}

\item{value}{valid value for the specified variable}
}
\description{
Setter function to set a variable of a simdesign object
}
\examples{
\dontrun{
setsim(nl, "simseeds") <- c(123, 456, 789)
}
}
