% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_init.R
\name{simdesign-class}
\alias{simdesign-class}
\alias{.initClasses}
\title{S4 class \code{"simdesign"}}
\usage{
.initClasses()
}
\description{
empty

empty

empty
}
\section{Slots}{

\describe{
\item{\code{simmethod}}{A character string for method identification}

\item{\code{siminput}}{Input parameter tibble for model simulations}

\item{\code{simobject}}{Store method object (e.g. sobol, morris, eFast)}

\item{\code{simseeds}}{A vector of model seeds}

\item{\code{simoutput}}{variable to attach simulation result tibbles}

\item{\code{expname}}{A character string defining the name of the experiment}

\item{\code{outpath}}{Path to a directory where experiment output will be stored}

\item{\code{repetition}}{The number of repetitions for each row of the simulation design input tibble}

\item{\code{tickmetrics}}{"true" runs defined metrics on each simulation tick. "false" runs metrics only after simulation is finished}

\item{\code{idsetup}}{character string or vector of character strings, defining the name of the NetLogo setup procedure}

\item{\code{idgo}}{character string or vector of character strings, defining the name of the NetLogo go procedure}

\item{\code{idfinal}}{character string or vector of character strings, defining the name of NetLogo procedures that should be run after the last tick}

\item{\code{idrunnum}}{character string, defining the name of a NetLogo global that should be used to parse the current siminputrow during model executions which can then be used for self-written output.}

\item{\code{runtime}}{number of model ticks that should be run for each simulation}

\item{\code{evalticks}}{number or vector of tick numbers defining when measurements are taken}

\item{\code{stopcond}}{a NetLogo reporter that reports TRUE/FALSE. If it reports TRUE the current simulation run is stopped (e.g. "not any? turtles")}

\item{\code{metrics}}{vector of strings defining valid NetLogo reporters that are taken as output measurements (e.g. c("count turtles", "count patches"))}

\item{\code{metrics.turtles}}{vector of strings defining valid turtles-own variables that are taken as output measurements (e.g. c("who", "pxcor", "pycor", "color"))}

\item{\code{metrics.patches}}{vector of strings defining valid patches-own variables that are taken as output measurements (e.g. c("pxcor", "pycor", "pcolor"))}

\item{\code{metrics.links}}{vector of strings defining valid links-own variables that are taken as output measurements (e.g. c("end1", "end2"))}

\item{\code{variables}}{a nested list of variables that are changed within a simulation design. The name of each sublist item has to be a valid global of the defined NetLogo model. Each list item consist of a min value, a max value, a step value and a qfun (e.g. list("paramA" = list(min=0, max=1, step=0.1, qfun="qunif")))}

\item{\code{constants}}{a list of constants that are kept constant within a simulation design. The name of each list item has to be a valid global of the defined NetLogo model (e.g. list("pNUM" = 12, "pLOGIC"="TRUE", "pSTRING"=""default""))}

\item{\code{simdesign}}{Holds a simdesign S4 class object}

\item{\code{nlversion}}{A character string defining the NetLogo version that is used}

\item{\code{nlpath}}{Path to the NetLogo main directory matching the defined version}

\item{\code{modelpath}}{Path to the NetLogo model file (*.nlogo) that is used for simulations}

\item{\code{jvmmem}}{Java virtual machine memory capacity in megabytes}

\item{\code{experiment}}{Holds a experiment S4 class object}
}}

