% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdesign_helper.R
\name{simdesign_ff}
\alias{simdesign_ff}
\title{Add a full-factorial simdesign to a nl object}
\usage{
simdesign_ff(nl, nseeds)
}
\arguments{
\item{nl}{nl object with a defined experiment}

\item{nseeds}{number of seeds for this simulation design}
}
\description{
Add a full-factorial simdesign to a nl object
}
\details{
This function creates a simdesign S4 class which can be added to a nl object by using the setter function simdesign(nl).
The full-factorial simdesign uses parameter ranges that are defined in the variables field of the experiment object within the nl object.
First, a vector is created for each parameter, based on the min, max and step value.
Then, a full-factorial matrix of all parameter combinations is created as input tibble for the simdesign.
Finally, the function reports a simdesign object.
}
\examples{
\dontrun{
# Example for Wolf Sheep Predation model from NetLogo models library:
nl@simdesign <- simdesign_ff(nl = nl, nseeds = 3)
}

}
