% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdesign_helper.R
\name{simdesign_simple}
\alias{simdesign_simple}
\title{Add a simple simdesign to a nl object}
\usage{
simdesign_simple(nl, nseeds)
}
\arguments{
\item{nl}{nl object with a defined experiment}

\item{nseeds}{number of seeds for this simulation design}
}
\description{
Add a simple simdesign to a nl object
}
\details{
This function creates a simdesign S4 class which can be added to a nl object by using the setter function simdesign(nl).
The simple simdesign only uses model parameters that are defined in the constants field of the experiment object within the nl object.
Thus, the resulting input tibble of the simdesign has only one run with constant parameterisations.
This can be useful to run one simulation with a specific parameterset.
Finally, the function reports a simdesign object.
}
\examples{
\dontrun{
# Example for Wolf Sheep Predation model from NetLogo models library:
nl@simdesign <- simdesign_simple(nl = nl, nseeds = 3)
}

}
