% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl.R
\name{util_cleanup}
\alias{util_cleanup}
\title{Delete temporary files}
\usage{
util_cleanup(nl, pattern)
}
\arguments{
\item{nl}{nl object}

\item{pattern}{defines file types to be deleted (".xml", ".csv" or "all")}
}
\description{
Delete temporary files
}
