% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_eval.R
\name{util_eval_constants}
\alias{util_eval_constants}
\title{Evaluate constants list of an experiment object}
\usage{
util_eval_constants(nl)
}
\arguments{
\item{nl}{nl object}
}
\description{
Evaluate constants list of an experiment object
}
\keyword{internal}
