% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_eval.R
\name{util_eval_experiment}
\alias{util_eval_experiment}
\title{Evaluate all slots of an experiment object}
\usage{
util_eval_experiment(nl)
}
\arguments{
\item{nl}{nl object}
}
\description{
Evaluate all slots of an experiment object
}
\details{
util_eval_experiment checks if the information stored within the experiment slots are valid.
}
\keyword{internal}
