% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl_dyn.R
\name{util_run_nl_dyn_ABCmcmc}
\alias{util_run_nl_dyn_ABCmcmc}
\title{ABCmcmc call simulations function}
\usage{
util_run_nl_dyn_ABCmcmc(
  nl,
  seed,
  cleanup.csv = TRUE,
  cleanup.xml = TRUE,
  cleanup.bat = TRUE
)
}
\arguments{
\item{nl}{nl object}

\item{seed}{current model seed}

\item{cleanup.csv}{TRUE/FALSE, if TRUE temporary created csv output files will be deleted after gathering results.}

\item{cleanup.xml}{TRUE/FALSE, if TRUE temporary created xml output files will be deleted after gathering results.}

\item{cleanup.bat}{TRUE/FALSE, if TRUE temporary created bat/sh output files will be deleted after gathering results.}
}
\description{
ABCmcmc call simulations function
}
\keyword{internal}
