% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl_dyn.R
\name{util_run_nl_dyn_GenAlg}
\alias{util_run_nl_dyn_GenAlg}
\title{Genetic Algorithm call simulations function}
\usage{
util_run_nl_dyn_GenAlg(nl, seed, threads = 1, ...)
}
\arguments{
\item{nl}{nl object}

\item{seed}{current model seed}

\item{threads}{number of NetLogo threads used for execution (handled via NetLogo).}

\item{cleanup.csv}{Deprecated. No longer has any effect.}

\item{cleanup.xml}{Deprecated. No longer has any effect.}

\item{cleanup.bat}{Deprecated. No longer has any effect.}
}
\description{
Genetic Algorithm call simulations function
}
\keyword{internal}
