% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl_dyn.R
\name{util_run_nl_dyn_GenAlg}
\alias{util_run_nl_dyn_GenAlg}
\title{Genetic Algorithm call simulations function}
\usage{
util_run_nl_dyn_GenAlg(nl, seed, cleanup)
}
\arguments{
\item{nl}{nl object}

\item{seed}{current model seed}

\item{cleanup}{indicate which filetypes should be deleted}
}
\description{
Genetic Algorithm call simulations function
}
\keyword{internal}
