% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl_dyn.R
\name{util_run_nl_dyn_GenAlg_fn}
\alias{util_run_nl_dyn_GenAlg_fn}
\title{Genetic Algorithm run simulation function}
\usage{
util_run_nl_dyn_GenAlg_fn(param, nl, evalcrit, seed, cleanup)
}
\arguments{
\item{param}{vector of model parameters, generated by GenSA function}

\item{nl}{nl object}

\item{evalcrit}{evaluation criterion for simulated annealing}

\item{seed}{current model seed}

\item{cleanup}{indicate which filetypes should be deleted}
}
\description{
Genetic Algorithm run simulation function
}
\keyword{internal}
