% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl_dyn.R
\name{util_run_nl_dyn_GenAlg_fn}
\alias{util_run_nl_dyn_GenAlg_fn}
\alias{util_run_nl_dyn_ABCmcmc_fn}
\title{Genetic Algorithm run simulation function}
\usage{
util_run_nl_dyn_GenAlg_fn(
  param,
  nl,
  evalcrit,
  seed,
  cleanup.csv,
  cleanup.xml,
  cleanup.bat
)

util_run_nl_dyn_ABCmcmc_fn(param)
}
\arguments{
\item{param}{vector of model parameters passed from ABC_mcmc function. If use_seeds = TRUE, the first element of this vector is a random seed}

\item{nl}{nl object}

\item{evalcrit}{evaluation criterion for simulated annealing}

\item{seed}{current model seed}

\item{cleanup.csv}{TRUE/FALSE, if TRUE temporary created csv output files will be deleted after gathering results.}

\item{cleanup.xml}{TRUE/FALSE, if TRUE temporary created xml output files will be deleted after gathering results.}

\item{cleanup.bat}{TRUE/FALSE, if TRUE temporary created bat/sh output files will be deleted after gathering results.}
}
\description{
Genetic Algorithm run simulation function

Genetic Algorithm run simulation function
}
\keyword{internal}
