% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl_dyn.R
\name{util_run_nl_dyn_GenSA}
\alias{util_run_nl_dyn_GenSA}
\title{Simulated Annealing call simulations function}
\usage{
util_run_nl_dyn_GenSA(nl, seed, threads = 1, ...)
}
\arguments{
\item{nl}{nl object}

\item{seed}{current model seed}

\item{threads}{number of NetLogo threads used for execution (handled via NetLogo).}
}
\description{
Simulated Annealing call simulations function
}
\keyword{internal}
