% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl_dyn.R
\name{util_run_nl_dyn_GenSA}
\alias{util_run_nl_dyn_GenSA}
\title{Simulated Annealing call simulations function}
\usage{
util_run_nl_dyn_GenSA(nl, seed, cleanup.csv, cleanup.xml, cleanup.bat)
}
\arguments{
\item{nl}{nl object}

\item{seed}{current model seed}

\item{cleanup.csv}{TRUE/FALSE, if TRUE temporary created csv output files will be deleted after gathering results.}

\item{cleanup.xml}{TRUE/FALSE, if TRUE temporary created xml output files will be deleted after gathering results.}

\item{cleanup.bat}{TRUE/FALSE, if TRUE temporary created bat/sh output files will be deleted after gathering results.}
}
\description{
Simulated Annealing call simulations function
}
\keyword{internal}
