% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl_dyn.R
\name{util_run_nl_dyn_GenSA_fn}
\alias{util_run_nl_dyn_GenSA_fn}
\title{Simulated Annealing run simulation function}
\usage{
util_run_nl_dyn_GenSA_fn(param, nl, evalcrit, seed, threads = 1, ...)
}
\arguments{
\item{param}{vector of model parameters, generated by GenSA function}

\item{nl}{nl object}

\item{evalcrit}{evaluation criterion for simulated annealing}

\item{seed}{current model seed}

\item{threads}{number of NetLogo threads used for execution (handled via NetLogo).}

\item{cleanup.csv}{Deprecated. No longer has any effect.}

\item{cleanup.xml}{Deprecated. No longer has any effect.}

\item{cleanup.bat}{Deprecated. No longer has any effect.}
}
\description{
Simulated Annealing run simulation function
}
\keyword{internal}
