% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_runnl_dyn.R
\name{util_run_nl_dyn_GenSA_fn}
\alias{util_run_nl_dyn_GenSA_fn}
\title{Simulated Annealing run simulation function}
\usage{
util_run_nl_dyn_GenSA_fn(param, nl, evalcrit, seed, cleanup.csv,
  cleanup.xml, cleanup.bat, silent)
}
\arguments{
\item{param}{vector of model parameters, generated by GenSA function}

\item{nl}{nl object}

\item{evalcrit}{evaluation criterion for simulated annealing}

\item{seed}{current model seed}

\item{cleanup.csv}{TRUE/FALSE, if TRUE temporary created csv output files will be deleted after gathering results.}

\item{cleanup.xml}{TRUE/FALSE, if TRUE temporary created xml output files will be deleted after gathering results.}

\item{cleanup.bat}{TRUE/FALSE, if TRUE temporary created bat/sh output files will be deleted after gathering results.}

\item{silent}{TRUE/FALSE, if FALSE prints the current seed and siminputrow after successful execution of a simulation (only for sequential execution).}
}
\description{
Simulated Annealing run simulation function
}
\keyword{internal}
