% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nl.R
\name{write_simoutput}
\alias{write_simoutput}
\title{Write attached NetLogo simulation output to file}
\usage{
write_simoutput(nl, outpath = NA)
}
\arguments{
\item{nl}{nl object}

\item{outpath}{optional path to directory where output is written

Write NetLogo simulation output to a csv file in the directory outpath of the nl object
Output has to be attached to the simdesign first with simoutput(nl) <- results
The outpath argument can be optionally used to write output to a different directory than the defined outpath of the nl object.}
}
\description{
Write attached NetLogo simulation output to file
}
\examples{

# Load nl object including output data from testdata
nl <- nl_lhs

# Write output to outpath directory
write_simoutput(nl, outpath=tempdir())

}
